net: separate rpc metrics from generic transport
The net::transport metrics managed a superset of metrics for all
specializations (e.g. rpc, kafka client). This commit separates them so
that the client transport specializations can manage their own metrics.

It's not pretty, but there is also only one user of it in the tree: the
rpc transport. With at least one more example the pattern could likely
be made more tasty.

The basic strategy is to have a transport probe inherit from the base
transport probe. The specialization owns the probe, and passes in a
reference to the base transport which will update the metrics common to
all transports.

When the specialization sets up metrics it collects its own and passes
them to the base transport metrics setup function which combines all of
them together and registers them.
