+++ src/v/net/client_probe.h
@@ -10,16 +10,16 @@
  */
 
 #pragma once
+
 #include "metrics/metrics.h"
-#include "model/metadata.h"
-#include "net/unresolved_address.h"
-#include "rpc/types.h"
 
 #include <seastar/core/metrics_registration.hh>
 
-#include <iosfwd>
-
 namespace net {
+
+/*
+ * Metrics probe that is common to all net::base_transport instances.
+ */
 class client_probe {
 public:
     client_probe() = default;
@@ -29,73 +29,32 @@ public:
     client_probe& operator=(client_probe&&) = delete;
     ~client_probe() = default;
 
-    void request() {
-        ++_requests;
-        ++_requests_pending;
-    }
-
-    void request_completed() {
-        ++_requests_completed;
-        --_requests_pending;
-    }
-
-    void request_timeout() {
-        ++_request_timeouts;
-        --_requests_pending;
-    }
-
-    void request_error() {
-        ++_request_errors;
-        --_requests_pending;
-    }
-
-    void add_bytes_sent(size_t sent) { _out_bytes += sent; }
-
-    void add_bytes_received(size_t recv) { _in_bytes += recv; }
-
     void connection_established() {
         ++_connects;
         ++_connections;
     }
 
-    void connection_closed() { --_connections; }
-
     void connection_error() { ++_connection_errors; }
 
-    void read_dispatch_error() { ++_read_dispatch_errors; }
+    void connection_closed() { --_connections; }
 
-    void header_corrupted() { ++_corrupted_headers; }
-
-    void client_correlation_error() { ++_client_correlation_errors; }
-
-    void server_correlation_error() { ++_server_correlation_errors; }
-
-    void waiting_for_available_memory() { ++_requests_blocked_memory; }
-
+    /*
+     * setup metrics
+     *
+     * name: name of the metrics group
+     * defs: additional metrics to register
+     */
     void setup_metrics(
-      metrics::internal_metric_groups& mgs,
-      const std::optional<rpc::connection_cache_label>& label,
-      const std::optional<model::node_id>& node_id,
-      const net::unresolved_address& target_addr);
+      std::string_view name,
+      const std::vector<ss::metrics::label_instance>& labels,
+      const std::vector<ss::metrics::label>& aggregate_labels,
+      std::vector<ss::metrics::metric_definition> defs);
 
-private:
-    uint64_t _requests = 0;
-    uint32_t _requests_pending = 0;
-    uint32_t _request_errors = 0;
-    uint64_t _request_timeouts = 0;
-    uint64_t _requests_completed = 0;
-    uint64_t _in_bytes = 0;
-    uint64_t _out_bytes = 0;
+protected:
     uint64_t _connects = 0;
     uint32_t _connections = 0;
     uint32_t _connection_errors = 0;
-    uint32_t _read_dispatch_errors = 0;
-    uint32_t _corrupted_headers = 0;
-    uint32_t _server_correlation_errors = 0;
-    uint32_t _client_correlation_errors = 0;
-    uint32_t _requests_blocked_memory = 0;
     metrics::internal_metric_groups _metrics;
-
-    friend std::ostream& operator<<(std::ostream& o, const client_probe& p);
 };
+
 }; // namespace net
