+++ src/v/rpc/transport.cc
@@ -66,7 +66,9 @@ transport::transport(
     &rpclog)
   , _memory(c.max_queued_bytes, "rpc/transport-mem")
   , _version(c.version)
-  , _default_version(c.version) {
+  , _default_version(c.version)
+  , _probe(std::make_unique<client_probe>()) {
+    set_probe(_probe.get());
     if (!c.disable_metrics) {
         setup_metrics(label, node_id);
     }
@@ -435,7 +437,28 @@ ss::future<> transport::dispatch(header h) {
 void transport::setup_metrics(
   const std::optional<connection_cache_label>& label,
   const std::optional<model::node_id>& node_id) {
-    _probe->setup_metrics(_metrics, label, node_id, server_address());
+    namespace sm = ss::metrics;
+    auto target = sm::label("target");
+    std::vector<sm::label_instance> labels = {target(
+      ssx::sformat("{}:{}", server_address().host(), server_address().port()))};
+    if (label) {
+        labels.push_back(sm::label("connection_cache_label")((*label)()));
+    }
+    std::vector<sm::label> aggregate_labels;
+    // Label the metrics for a given server with the node ID so Seastar can
+    // differentiate between them, in case multiple node IDs start at the same
+    // address (e.g. in an ungraceful decommission). Aggregate on node ID so
+    // the user is presented metrics for each server regardless of node ID.
+    if (node_id) {
+        auto node_id_label = sm::label("node_id");
+        labels.push_back(node_id_label(*node_id));
+        aggregate_labels.push_back(node_id_label);
+    }
+    _probe->setup_metrics(
+      "rpc_client",
+      labels,
+      aggregate_labels,
+      _probe->defs(labels, aggregate_labels));
 }
 
 timing_info* transport::get_timing(uint32_t correlation) {
@@ -444,7 +467,8 @@ timing_info* transport::get_timing(uint32_t correlation) {
 }
 
 transport::~transport() {
-    vlog(rpclog.debug, "RPC Client to {} probes: {}", server_address(), _probe);
+    vlog(
+      rpclog.debug, "RPC Client to {} probes: {}", server_address(), *_probe);
     vassert(
       !is_valid(),
       "connection '{}' is still valid. must call stop() before "
@@ -461,4 +485,115 @@ std::ostream& operator<<(std::ostream& o, const transport& t) {
       t._correlation_idx);
     return o;
 }
+
+std::vector<ss::metrics::metric_definition> client_probe::defs(
+  const std::vector<ss::metrics::label_instance>& labels,
+  const std::vector<ss::metrics::label>& aggregate_labels) {
+    namespace sm = ss::metrics;
+    std::vector<sm::metric_definition> ret;
+
+    ret.emplace_back(sm::make_counter(
+                       "requests",
+                       [this] { return _requests; },
+                       sm::description("Number of requests"),
+                       labels)
+                       .aggregate(aggregate_labels));
+
+    ret.emplace_back(sm::make_gauge(
+                       "requests_pending",
+                       [this] { return _requests_pending; },
+                       sm::description("Number of requests pending"),
+                       labels)
+                       .aggregate(aggregate_labels));
+
+    ret.emplace_back(sm::make_counter(
+                       "request_errors",
+                       [this] { return _request_errors; },
+                       sm::description("Number or requests errors"),
+                       labels)
+                       .aggregate(aggregate_labels));
+
+    ret.emplace_back(sm::make_counter(
+                       "request_timeouts",
+                       [this] { return _request_timeouts; },
+                       sm::description("Number or requests timeouts"),
+                       labels)
+                       .aggregate(aggregate_labels));
+
+    ret.emplace_back(
+      sm::make_total_bytes(
+        "in_bytes",
+        [this] { return _out_bytes; },
+        sm::description("Total number of bytes sent (including headers)"),
+        labels)
+        .aggregate(aggregate_labels));
+
+    ret.emplace_back(sm::make_total_bytes(
+                       "out_bytes",
+                       [this] { return _in_bytes; },
+                       sm::description("Total number of bytes received"),
+                       labels)
+                       .aggregate(aggregate_labels));
+    ret.emplace_back(
+      sm::make_counter(
+        "read_dispatch_errors",
+        [this] { return _read_dispatch_errors; },
+        sm::description("Number of errors while dispatching responses"),
+        labels)
+        .aggregate(aggregate_labels));
+
+    ret.emplace_back(
+      sm::make_counter(
+        "corrupted_headers",
+        [this] { return _corrupted_headers; },
+        sm::description("Number of responses with corrupted headers"),
+        labels)
+        .aggregate(aggregate_labels));
+
+    ret.emplace_back(
+      sm::make_counter(
+        "server_correlation_errors",
+        [this] { return _server_correlation_errors; },
+        sm::description("Number of responses with wrong correlation id"),
+        labels)
+        .aggregate(aggregate_labels));
+
+    ret.emplace_back(
+      sm::make_counter(
+        "client_correlation_errors",
+        [this] { return _client_correlation_errors; },
+        sm::description("Number of errors in client correlation id"),
+        labels)
+        .aggregate(aggregate_labels));
+
+    ret.emplace_back(
+      sm::make_counter(
+        "requests_blocked_memory",
+        [this] { return _requests_blocked_memory; },
+        sm::description("Number of requests that are blocked because"
+                        " of insufficient memory"),
+        labels)
+        .aggregate(aggregate_labels));
+
+    return ret;
+}
+
+std::ostream& operator<<(std::ostream& o, const client_probe& p) {
+    o << "{"
+      << " requests_sent: " << p._requests
+      << ", requests_pending: " << p._requests_pending
+      << ", requests_completed: " << p._requests_completed
+      << ", request_errors: " << p._request_errors
+      << ", request_timeouts: " << p._request_timeouts
+      << ", in_bytes: " << p._in_bytes << ", out_bytes: " << p._out_bytes
+      << ", connects: " << p._connects << ", connections: " << p._connections
+      << ", connection_errors: " << p._connection_errors
+      << ", read_dispatch_errors: " << p._read_dispatch_errors
+      << ", corrupted_headers: " << p._corrupted_headers
+      << ", server_correlation_errors: " << p._server_correlation_errors
+      << ", client_correlation_errors: " << p._client_correlation_errors
+      << ", requests_blocked_memory: " << p._requests_blocked_memory << " }";
+    return o;
+}
+
 } // namespace rpc
